+++ src/endstone_example/listener.py
@@ -1,7 +1,5 @@
-import datetime
-
 from endstone import ColorFormat
-from endstone.event import EventPriority, PlayerJoinEvent, PlayerQuitEvent, ServerListPingEvent, event_handler
+from endstone.event import PlayerJoinEvent, PlayerQuitEvent, event_handler
 from endstone.plugin import Plugin
 
 
@@ -9,24 +7,12 @@ class ExampleListener:
     def __init__(self, plugin: Plugin) -> None:
         self._plugin = plugin
 
-    @event_handler(priority=EventPriority.HIGHEST)
-    def on_server_list_ping(self, event: ServerListPingEvent) -> None:
-        event.motd = ColorFormat.BOLD + ColorFormat.AQUA + datetime.datetime.now().strftime("%c")
-        addr = event.address
-        event.level_name = f"Your IP is {ColorFormat.YELLOW}{addr.hostname}:{addr.port}{ColorFormat.RESET}"
-
     @event_handler
     def on_player_join(self, event: PlayerJoinEvent) -> None:
         player = event.player
-        self._plugin.logger.info(
-            ColorFormat.YELLOW + f"{player.name}[/{player.address}] joined the game with UUID {player.unique_id}"
-        )
-
-        # example of explicitly removing one's permission of using /me command
-        player.add_attachment(self._plugin, "minecraft.command.me", False)
-        player.update_commands()  # don't forget to resend the commands
+        event.join_message = f"{ColorFormat.YELLOW}{player.name} joined the game"
+        self._plugin.logger.info(f"{player.name} joined from {player.address}")
 
     @event_handler
     def on_player_quit(self, event: PlayerQuitEvent) -> None:
-        player = event.player
-        self._plugin.logger.info(ColorFormat.YELLOW + f"{player.name}[/{player.address}] left the game.")
+        event.quit_message = f"{ColorFormat.YELLOW}{event.player.name} left the game"